feat: Improve prefer_sliver_prefix to prefer_to_include_sliver_in_name
