--- packages/altive_lints/lint_test/lints/prefer_sliver_prefix.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-
-// expect_lint: prefer_sliver_prefix
-class MyWidget extends StatelessWidget {
-  const MyWidget({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return SliverList.builder(
-      itemCount: 10,
-      itemBuilder: (context, index) {
-        return const Placeholder();
-      },
-    );
-  }
-}
