+++ packages/altive_lints/lint_test/lints/prefer_to_include_sliver_in_name.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+
+// expect_lint: prefer_to_include_sliver_in_name
+class MyWidget extends StatelessWidget {
+  const MyWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return SliverList.builder(
+      itemCount: 10,
+      itemBuilder: (context, index) {
+        return const Placeholder();
+      },
+    );
+  }
+}
+
+// expect_lint: prefer_to_include_sliver_in_name
+class MyWidget2 extends StatelessWidget {
+  const MyWidget2({super.key, this.maxCount = 10});
+
+  const MyWidget2.small({super.key, this.maxCount = 5});
+
+  final int maxCount;
+
+  @override
+  Widget build(BuildContext context) {
+    return SliverList.builder(
+      itemCount: maxCount,
+      itemBuilder: (context, index) {
+        return const Placeholder();
+      },
+    );
+  }
+}
